+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  application virtual server
 * [improvement] Add `add` alias for `shelly config create`
 * [improvement] Add `new` and `create` aliases for `shelly user add`
+* [improvement] Add `new` alias for `shelly backup create`
 
 ## 0.4.24 / 2014-02-25